\name{test2}
\alias{test2}
\title{
	test2
}
\description{
	This function is a test.
}
\usage{
	test2(x, ...)
}
\arguments{
  \item{x}{
	a numerical variable.}
}

\details{
  See the package's vignette for details on test.
}
\value{
  The returned value is a numerical variable.
}
\references{
	XX
}
\author{Franck Jabot}
\examples{
 \dontrun{
    ##### EXAMPLE 1 #####
    #####################
    x=2
    y=test2(x)
    y
 }%dontrun
}
